requirements.txt
numpy>=1.22
pandas>=1.5
pyarrow>=12
fastparquet>=2023.8.0
requests>=2.31
tqdm>=4.66
scikit-learn>=1.2
python-dateutil>=2.8
pyyaml>=6.0
scipy>=1.9
